requirements.txt
discord.py
psutil
timeago==1.0.13
python-dotenv
discord-pretty-help
python-dotenv